+++ package.json
@@ -21,8 +21,7 @@
     "react-toastify": "^10.0.0",
     "recoil-persist": "^5.1.0",
     "slick-carousel": "^1.8.1",
-    "styled-reset": "^4.5.2",
-    "vite-plugin-favicon": "^1.0.8"
+    "styled-reset": "^4.5.2"
   },
   "devDependencies": {
     "@supabase/supabase-js": "^2.39.2",
