+++ src/components/Common/Layout.tsx
@@ -1,5 +1,5 @@
 import {Outlet} from 'react-router-dom';
-import Nav from './Nav';
+import Nav from './Nav/index';
 import Footer from './Footer';
 
 const Layout = () => {
